Alpha 0.1
This is an alpha 0.1 version of the game.
11 Enemies will spawn at a random position.

Enemies will get deleted if they go more than 100 units from the center and no new enemies will spawn after.
This should be fixed in the next version.

 You can consume asteroids which will give you 1 mass and you will grow.

Also hoping to implement enemies spawning in random masses for the next version too.